Move the query request and response details to modules (#57)
