+++ modules/src/error.rs
@@ -1,0 +1,17 @@
+use anomaly::{BoxError, Context};
+use thiserror::Error;
+
+pub type Error = anomaly::Error<Kind>;
+
+#[derive(Clone, Debug, Error)]
+pub enum Kind {
+    /// RPC error (typically raised by the RPC client or the RPC requester)
+    #[error("RPC error")]
+    Rpc,
+}
+
+impl Kind {
+    pub fn context(self, source: impl Into<BoxError>) -> Context<Self> {
+        Context::new(self, Some(source.into()))
+    }
+}
