+++ relayer/cli/src/commands/query/client.rs
@@ -2,8 +2,7 @@ use crate::prelude::*;
 
 use abscissa_core::{Command, Options, Runnable};
 use relayer::config::{ChainConfig, Config};
-use relayer::query::client_consensus_state::query_client_consensus_state;
-use relayer::query::client_state::query_client_full_state;
+use relayer::query::client::{query_client_consensus_state, query_client_full_state};
 
 use relayer_modules::ics24_host::client::ClientId;
